Add files via upload
It is a rock-paper-scissor-game in this I have used the random module and its sub methods like random.choice() that enables computer to choose Rock or Paper or Scissor.
As we know paper wins over rock and scissor wins over paper and paper beats rock. Based on computer choice user may either wins or lose or sometime there will be a tie!. Until user quits the game it will continuously loops.